+++ java/com/controller/admin/AdminRoomImageDeleteController.java
@@ -2,6 +2,7 @@ package com.controller.admin;
 
 import java.io.File;
 
+import com.config.UploadPath;
 import com.controller.Action;
 import com.dao.AdminDAO;
 
@@ -19,12 +20,19 @@ public class AdminRoomImageDeleteController implements Action {
 		AdminDAO dao = new AdminDAO(request.getServletContext());
 		dao.deleteRoomImage(image_path);
 
-		String realPath = request.getServletContext().getRealPath(image_path);
+		// 파일 이름 추출
+		String fileName = new File(image_path).getName();
 
-		File file = new File(realPath);
-		if (file.exists()) {
+		// 실제 업로드 폴더
+		String uploadPath = UploadPath.ROOM;
+
+		File file = new File(uploadPath, fileName);
+
+		if (file.exists() && file.isFile()) {
 			file.delete();
 		}
+
 		return "redirect:/admin/roomDetail.do?room_id=" + room_id;
 	}
+
 }
